factor out local_robot tests relying on tflite, flag
